Allow boot-deploy config be customized by distros, relocate to /usr/share (MR 30)
The configuration from boot-deploy comes from the deviceinfo package.
Adding another file where any kind of configuration can be added increases
the risks of making mistakes and start pushing any kind of configuration
to it. Since the file was born for other distros to be able to configure
boot-deploy[1][2], let's reflect that in the file name. In the process,
move the file to /usr/share from /etc, since it is not a user-owned (host-
specific) but a distro-owned (should not be modified) file.

Also, update the documentation corresponding to the file, since both the
file and all its variables are boot-deploy requirements.

[1] https://gitlab.com/postmarketOS/boot-deploy/-/merge_requests/18#note_1177011687
[2] https://gitlab.com/postmarketOS/boot-deploy/-/merge_requests/18#note_1177011689