Use "local" for variable scoping
This introduces using "local" in the scripts to scope variables
locally. "local" is not defined in POSIX, so this breaks compatibility.
At least in Alpine, which uses ash as it's "POSIX shell", this is
supported.

We re-use a lot of variable names (e.g. "$file") and this can lead to
some interesting bugs if we're not careful, so it seems worth breaking
POSIX compatibility in order to get real variable scoping to avoid this
class of bugs.