Add validation for deviceinfo variables
It's expected that some values are lowercase, per the deviceinfo
spec. This is because some values are used in boot-deploy for string
comparison, and sh can't do case-insensitive comparisons, so "True" !=
"true" != "TrUe" and so on. This adds a check to validate that
deviceinfo vars have lower case values, at least the ones that are used
for string comparison in the script.

Output looks like:
    ERROR: variable should have a lowercase value: deviceinfo_append_dtb
    ERROR: variable should have a lowercase value: deviceinfo_arch
    For more information, see: https://wiki.postmarketos.org/wiki/Deviceinfo_reference#Case_sensitivity

I opted to just print every failing var so it's more helpful.