Add support for systemd-boot
This adds support for systemd-boot, by combining it with support for
gummiboot (the gummiboot project turned into "systemd-boot"...). This
also installs all available bootloaders under the relevant systemd-boot
and gummiboot directories. That's useful because it allows something
outside of boot-deploy to control what boot-deploy installs, without
having to add a bunch of new logic + new deviceinfo var for it.

The `generate_gummiboot` var was renamed to `generate_systemd_boot`, one
turned into the other, and the internal logic in boot-deploy will now
search dirs that both gummiboot and systemd-boot are installed in. It
prefers systemd-boot over gummiboot, in case both options are located on
the system.

deviceinfo changes:

- ADDED: deviceinfo_generate_systemd_boot
        - triggers installing support for systemd-boot

- DROP: deviceinfo_generate_gummiboot
        - Use deviceinfo_generate_systemd_boot to trigger gummiboot
          support