find_dtb: remove newlines from output
Multiple dtbs can be returned if globbing is used, previously this
included newlines in the output... I'm assuming that's not desired for
any usage of find_dtb, so this removes the newline.

This also mentions globbing in the function description, and drops the
comment about how find_dtb only returns the first dtb it finds... which
is not true.