systemd: prefer vmlinuz.efi over vmlinuz (MR 47)
For some kernels in postmarketOS it will be useful to support
systemd-booting (via a chainloaded bootloader like u-boot) as well as
supporting the existing bootloader which only supports a GZIP compressed
kernel image.

To enable this usecase, check to see if a version of the kernel file
with the `.efi` extension exists and prioritise using it for
systemd-boot.

Signed-off-by: Caleb Connolly <caleb@connolly.tech>