copy target dtb to a temp directory for qcom and sprd dt.img types (MR 57)
The dtbTool and dtbTool-sprd programs require the dtb directory as an
argument. These programs tend to combine multiple devicetrees. We can't
determine for whether the directory has devicetree blobs for a single
device, or multiple devices.

It's safer to copy the target dtb over to a temporary directory, and
pass that directory as an argument to dtbTool.