Add support for generating compressed depthcharge images (MR 65)
Some chromebooks have a small kernel partition which barely fits the
depthcharge image, and in certain cases is too small for the depthcharge
image. Add support for compressing the kernel in the depthcharge image
which fixes this issue. Not every chromebook supports compressed kernels,
so a deviceinfo variable is added to specify the compression type
(either lzma or lz4). If the variable is unset, compression will not be used.