boot-deploy-functions: extlinux: support for dtb overlays
To allow more flexibility in managing device variations with
different peripherals, we canuse DTB overlays. Bootloader should
apply them, if it can detect and identify device variation.
Or if it is specified in its config.

Both U-Boot and lk2nd support DTB overlays for extlinux.

Add support for generating fdtoverlays="/... /..." line in
extlinux.conf if there is the "deviceinfo_dtb_overlays"
variable specified in deviceinfo.

Signed-off-by: Alexey Minnekhanov <alexeymin@postmarketos.org>