ci: use .shellcheckrc to manage exclusions
The advantage with .shellcheckrc is that the IDE LSP using shellcheck
can also read it. This reduces the innumerable amount of inline warnings
displayed. In other words, bliss for the developer.

This also drops the exception for SC1008, all scripts have the sh
shebang, and we should add it for new scripts if we forget (in which
case this test will fail).

Co-authored-by: Clayton Craft <clayton@craftyguy.net>
Part-of: https://gitlab.postmarketos.org/postmarketOS/boot-deploy/-/merge_requests/86
[ci:skip-build]: already built successfully in CI