Remove boot image header v2 spaghetti
Some devices that use use boot image header v2 also require the setting
of bootimg_custom_args. Currently the --dtb and --dtb_offset parameters
are silently not added to the mkbootimg invocation, leading to failure
to include the dtb in the boot image.

Currently in postmarketOS pmaports there is one device that has both
header_version="2" AND bootimg_custom_args set, which is the Fairphone
(Gen. 6), and it's breaking boot.img generation there. I didn't dig into
if at earlier times (e.g. when that code was introduced in July 2022)
there were occurances in pmaports where this broke but now this should
be safe to do.

Part-of: https://gitlab.postmarketos.org/postmarketOS/boot-deploy/-/merge_requests/94
[ci:skip-build]: already built successfully in CI